producer/requirements.txt
Faker==4.0.3
kafka-python==2.0.2